refactor(): * moved modules around / renamed / restructured ui * split huge 'main' view into smaller functions
